improve the mlflow.data code organization
