+++ mlflow/data/digest_utils.py
@@ -1,7 +1,5 @@
 from typing import Any, List
 
-from packaging.version import Version
-
 from mlflow.exceptions import MlflowException
 from mlflow.protos.databricks_pb2 import INVALID_PARAMETER_VALUE
 from mlflow.utils import insecure_hash
@@ -17,7 +15,6 @@ def compute_pandas_digest(df) -> str:
 
     Returns:
         A string digest.
-
     """
     import numpy as np
     import pandas as pd
@@ -50,7 +47,6 @@ def compute_numpy_digest(features, targets=None) -> str:
 
     Returns:
         A string digest.
-
     """
     import numpy as np
     import pandas as pd
@@ -84,82 +80,6 @@ def compute_numpy_digest(features, targets=None) -> str:
     return get_normalized_md5_digest(hashable_elements)
 
 
-def compute_tensorflow_dataset_digest(dataset, targets=None) -> str:
-    """Computes a digest for the given Tensorflow dataset.
-
-    Args:
-        dataset: A Tensorflow dataset.
-
-    Returns:
-        A string digest.
-
-    """
-    import numpy as np
-    import pandas as pd
-    import tensorflow as tf
-
-    hashable_elements = []
-
-    def hash_tf_dataset_iterator_element(element):
-        if element is None:
-            return
-        flat_element = tf.nest.flatten(element)
-        flattened_array = np.concatenate([x.flatten() for x in flat_element])
-        trimmed_array = flattened_array[0:MAX_ROWS]
-        try:
-            hashable_elements.append(pd.util.hash_array(trimmed_array))
-        except TypeError:
-            hashable_elements.append(np.int64(trimmed_array.size))
-
-    for element in dataset.as_numpy_iterator():
-        hash_tf_dataset_iterator_element(element)
-    if targets is not None:
-        for element in targets.as_numpy_iterator():
-            hash_tf_dataset_iterator_element(element)
-
-    return get_normalized_md5_digest(hashable_elements)
-
-
-def compute_tensor_digest(tensor_data, tensor_targets) -> str:
-    """Computes a digest for the given Tensorflow tensor.
-
-    Args:
-        tensor: A Tensorflow tensor.
-
-    Returns:
-        A string digest.
-
-    """
-    if tensor_targets is None:
-        return compute_numpy_digest(tensor_data.numpy())
-    else:
-        return compute_numpy_digest(tensor_data.numpy(), tensor_targets.numpy())
-
-
-def compute_spark_df_digest(df) -> str:
-    """Computes a digest for the given Spark DataFrame. Retrieve a semantic hash of the
-    DataFrame's logical plan, which is much more efficient and deterministic than hashing
-    DataFrame records
-
-    Args:
-        df: A Spark DataFrame.
-
-    Returns:
-        A string digest.
-
-    """
-
-    import numpy as np
-    import pyspark
-
-    # Spark 3.1.0+ has a semanticHash() method on DataFrame
-    if Version(pyspark.__version__) >= Version("3.1.0"):
-        semantic_hash = df.semanticHash()
-    else:
-        semantic_hash = df._jdf.queryExecution().analyzed().semanticHash()
-    return get_normalized_md5_digest([np.int64(semantic_hash)])
-
-
 def get_normalized_md5_digest(elements: List[Any]) -> str:
     """Computes a normalized digest for a list of hashable elements.
 
@@ -168,7 +88,6 @@ def get_normalized_md5_digest(elements: List[Any]) -> str:
 
     Returns:
         An 8-character, truncated md5 digest.
-
     """
 
     if not elements:
